requirements.txt
aiogram==3.5
python-dotenv>=1.0
aiohttp>=3.8
aiofiles~=23.2.1
magic-filter>=1.0.12,<1.1
pydantic<2.18
typing-extensions<=5.0,>=4.7.0